Relax PixelMap input requirements

+++ lib/DisplayPatternLib/src/DisplayPatterns/CenterOutDisplayPattern.h
@@ -24,8 +24,8 @@ class CenterOutDisplayPattern : public DisplayPattern {
 
     private:
         CenterOutOrientation m_orientation{CenterOutOrientation::Horizontal};
-        uint16_t m_centerRow;
-        uint16_t m_centerColumn;
+        int m_centerRow;
+        int m_centerColumn;
 
         void updateHorizontal(PixelMap* pixelMap, uint32_t newColor);
         void updateVertical(PixelMap* pixelMap, uint32_t newColor);
